http/tests/cache/disk-cache/disk-cache-validation-back-navigation-policy.html failing on Mavericks wk2 debug https://bugs.webkit.org/show_bug.cgi?id=143258
The test relies on resource priorities. Those don't roundtrip when serializing via NSURLRequest on Mavericks due
to ResourceRequest::resourcePrioritiesEnabled() returning false.

It is bit surprising it works on release but not on debug. I suspect some debug code forces NSURLRequest construction
and so we end up on the roundtrip code path.

* platform/mac-mavericks/TestExpectations: Added.

    Skip the test on Mavericks.



Identifier: 161281@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182181 268f45cc-cd09-0410-ab3c-d52691b4dbfc