[Win] Bots complain about missing /etc/catalog https://bugs.webkit.org/show_bug.cgi?id=143484
Reviewed by Chris Dumez.

The libxml2 build on Mac, iOS, and Windows does not provide the set
of default catalog files. We do not actually need these for our
purposes, but we get warnings and some bot flakiness because they
are not present.

We can avoid both problems the same way the Mac and iOS ports do,
by setting the XML_CATALOG_FILES environment variable to 'empty string',
which avoid the warnings and bot problems.

* Scripts/webkitpy/port/win.py:
(WinPort.setup_environ_for_server): Added.


Identifier: 161505@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182481 268f45cc-cd09-0410-ab3c-d52691b4dbfc